4.1.1
Worked on calculatePointValues() method and discovered an error in the vector class that creates a vector from two points. It is suspected the pi/4 angles that the algorithm gives to 0-magnitude components is one of the key things throwing off the updatePoints() method. (the problem may be able to be fixed by altering the interpretation of the point-angles in the updatePoints() method)